Refactor runContainer in docker-client
Refactored the runContainer method, splitting out the portion of it
which awaits the container in a separate method. This allows the
library user to have control over whether they block until the container
exits and also provides a mechanism to get the containerId in cases
where the container exits with an error code.