Update DockerTask interpeter hashing
Updated how DockerTask file outputs are copied to the CAS
so that the hashing method is called on the file content
instead of some arbitrary inputs (#67). This now writes
files to a temporary directory, computes the hash on that,
then copies it to the CAS. To support this 2-step operation
I also had to update `moveDirectoryContent` to copy files
in cases where the temp directory resides on a separate disk
from the CAS (e.g. this can happen when running in a nix-shell).